Generate baseline dates vector

Function to generate baseline dates given an end date and test length, plus optional guard, and length of baseline

Usage

get_baseline_dates(end_date, test_length, baseline_length, guard = 0)

Arguments

end_date

End date of the test interval

test_length

(integer) length of the test interval in days

baseline_length

(integer) length of baseline period in days

guard

(integer) default = 0; buffer between end of baseline and start of test interval

Value

vector of dates

Examples

get_baseline_dates(
  end_date = "2025-01-01",
  test_length = 10,
  baseline_length = 90
)
#>  [1] "2024-09-24" "2024-09-25" "2024-09-26" "2024-09-27" "2024-09-28"
#>  [6] "2024-09-29" "2024-09-30" "2024-10-01" "2024-10-02" "2024-10-03"
#> [11] "2024-10-04" "2024-10-05" "2024-10-06" "2024-10-07" "2024-10-08"
#> [16] "2024-10-09" "2024-10-10" "2024-10-11" "2024-10-12" "2024-10-13"
#> [21] "2024-10-14" "2024-10-15" "2024-10-16" "2024-10-17" "2024-10-18"
#> [26] "2024-10-19" "2024-10-20" "2024-10-21" "2024-10-22" "2024-10-23"
#> [31] "2024-10-24" "2024-10-25" "2024-10-26" "2024-10-27" "2024-10-28"
#> [36] "2024-10-29" "2024-10-30" "2024-10-31" "2024-11-01" "2024-11-02"
#> [41] "2024-11-03" "2024-11-04" "2024-11-05" "2024-11-06" "2024-11-07"
#> [46] "2024-11-08" "2024-11-09" "2024-11-10" "2024-11-11" "2024-11-12"
#> [51] "2024-11-13" "2024-11-14" "2024-11-15" "2024-11-16" "2024-11-17"
#> [56] "2024-11-18" "2024-11-19" "2024-11-20" "2024-11-21" "2024-11-22"
#> [61] "2024-11-23" "2024-11-24" "2024-11-25" "2024-11-26" "2024-11-27"
#> [66] "2024-11-28" "2024-11-29" "2024-11-30" "2024-12-01" "2024-12-02"
#> [71] "2024-12-03" "2024-12-04" "2024-12-05" "2024-12-06" "2024-12-07"
#> [76] "2024-12-08" "2024-12-09" "2024-12-10" "2024-12-11" "2024-12-12"
#> [81] "2024-12-13" "2024-12-14" "2024-12-15" "2024-12-16" "2024-12-17"
#> [86] "2024-12-18" "2024-12-19" "2024-12-20" "2024-12-21" "2024-12-22"